move ring buffer type to op-service
